requirements.txt
bilibili_api_python>=17.4.0
colorlog>=6.9.0
python-telegram-bot~=21.3
openai>=1.106.1
Requests~=2.32.4
tavily_python~=0.7.21
uuid~=1.30
websockets>=10.0
PyJWT>=2.12.1
fastapi>=0.124.0
uvicorn>=0.35.0
pydantic>=2.11.7
psutil>=5.9.0
fastmcp==2.14.4
httpx>=0.28.1
watchfiles>=1.1.1